+++ tests/resources/go/asyncapi-skip-unmarshal/entities.go
@@ -341,6 +341,9 @@ type marshalReference Reference
 
 // MarshalJSON encodes JSON.
 func (i Reference) MarshalJSON() ([]byte, error) {
+	if len(i.AdditionalProperties) == 0 {
+		return json.Marshal(marshalReference(i))
+	}
 	return marshalUnion(marshalReference(i), i.AdditionalProperties)
 }