+++ tests/resources/go/asyncapi2/entities.go
@@ -663,6 +663,9 @@ func (i *ServerBindingsObject) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON encodes JSON.
 func (i ServerBindingsObject) MarshalJSON() ([]byte, error) {
+	if len(i.AdditionalProperties) == 0 {
+		return json.Marshal(marshalServerBindingsObject(i))
+	}
 	return marshalUnion(marshalServerBindingsObject(i), i.AdditionalProperties)
 }
 
@@ -961,6 +964,9 @@ func (i *Reference) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON encodes JSON.
 func (i Reference) MarshalJSON() ([]byte, error) {
+	if len(i.AdditionalProperties) == 0 {
+		return json.Marshal(marshalReference(i))
+	}
 	return marshalUnion(marshalReference(i), i.AdditionalProperties)
 }
 
@@ -1356,6 +1362,9 @@ func (i *OperationBindingsObject) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON encodes JSON.
 func (i OperationBindingsObject) MarshalJSON() ([]byte, error) {
+	if len(i.AdditionalProperties) == 0 {
+		return json.Marshal(marshalOperationBindingsObject(i))
+	}
 	return marshalUnion(marshalOperationBindingsObject(i), i.AdditionalProperties)
 }
 
@@ -1788,6 +1797,9 @@ func (i *MessageBindingsObject) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON encodes JSON.
 func (i MessageBindingsObject) MarshalJSON() ([]byte, error) {
+	if len(i.AdditionalProperties) == 0 {
+		return json.Marshal(marshalMessageBindingsObject(i))
+	}
 	return marshalUnion(marshalMessageBindingsObject(i), i.AdditionalProperties)
 }
 
@@ -2209,6 +2221,9 @@ func (i *ChannelBindingsObject) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON encodes JSON.
 func (i ChannelBindingsObject) MarshalJSON() ([]byte, error) {
+	if len(i.AdditionalProperties) == 0 {
+		return json.Marshal(marshalChannelBindingsObject(i))
+	}
 	return marshalUnion(marshalChannelBindingsObject(i), i.AdditionalProperties)
 }
 
@@ -2289,6 +2304,9 @@ func (i *Exchange) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON encodes JSON.
 func (i Exchange) MarshalJSON() ([]byte, error) {
+	if len(i.AdditionalProperties) == 0 {
+		return json.Marshal(marshalExchange(i))
+	}
 	return marshalUnion(marshalExchange(i), i.AdditionalProperties)
 }
 
@@ -2356,6 +2374,9 @@ func (i *Queue) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON encodes JSON.
 func (i Queue) MarshalJSON() ([]byte, error) {
+	if len(i.AdditionalProperties) == 0 {
+		return json.Marshal(marshalQueue(i))
+	}
 	return marshalUnion(marshalQueue(i), i.AdditionalProperties)
 }