+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.22] - 2020-01-26
+
+### Added
+- Early return in `MarshalJSON` if there are no additional properties to improve performance.
+- JSON Schema maker library updated.
+
 ## [0.4.21] - 2020-01-24
 
 ### Fixed